feat: "비밀번호 입력 시 수정, 삭제"
비밀번호를 요청값으로 받아 저장된 비밀번호와 일치할 경우 수정, 삭제 되도록 기능을 구현하였습니다.

+++ src/main/java/com/example/schedule/service/ScheduleServiceImpl.java
@@ -60,12 +60,16 @@ public class ScheduleServiceImpl implements ScheduleService {
 
     @Override
     public ScheduleResponseDto updateSchedule(
-            Long id, String name, long password, String title, String content, String modified_at
+            long id, String name, long password, String title, String content, String modified_at
     ) {
         Schedule schedule = scheduleRepository.findScheduleById(id);
 
         if (schedule == null) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Does not exist id = " + id);
+        }
+
+        if (password != schedule.getPassword()) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Wrong password");
         }
 
         schedule.updateSchedule(name, title, content, modified_at);
@@ -74,12 +78,16 @@ public class ScheduleServiceImpl implements ScheduleService {
     }
 
     @Override
-    public void deleteSchedule(long id) {
+    public void deleteSchedule(long id, long password) {
 
         Schedule schedule = scheduleRepository.findScheduleById(id);
 
         if (schedule == null) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Does not exist id = " + id);
+        }
+
+        if (password != schedule.getPassword()) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Wrong password");
         }
 
         scheduleRepository.deleteSchedule(id);
